requirements.txt
# Core API
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
SQLAlchemy==2.0.36
alembic==1.14.0
mysqlclient==2.2.4

# Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Validation & Settings
pydantic==2.10.4
pydantic-settings==2.6.1
email-validator==2.2.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.19

starlette-context==0.3.6
cachetools==5.3.3
requests==2.31.0
Pillow==11.1.0
PyMuPDF==1.25.3
pillow-heif==0.21.0
reportlab==4.4.1
